Replace unsound attribute offset checks
Dereferencing a null pointer into a reference is always unsound. Using a
null pointer and even other techniques to determine attribute offsets
has already cause issues even in crates that are fully dedicated to it.

However, we are fortunate here as the structs in question are crafted
out of only basic integer types. Thus, it is safe to create an instances
of them that is fully zeroed with little code. Then the actual reference
to member access is fully safe and sound. An exception are only the
fields of unions which are inherently unsafe to access and using them to
create a reference is no exception.