Hide the unsafe to_str function
The implementation makes the assumption that the bytes making up the
C-string are both terminated and valid. This is wrong in general but
the zero-termination seems permissible as it only trusts the kernel's
correct implementation of POSIX.

The correct encoding trust is a different beast, there might be an
unterminated multi-byte character within the string.