redbpf, redbpf-probes: split user/kernel space MapData
redbpf_probes::xdp::MapData now includes only the kernel space API. The
user space API is now in redbpf::xdp::MapData.

The struct layout of MapData is now copy/pasted between the two modules.
It can't be shared because:

- redbpf can't depend on redbpf-probes, since redbpf-probes has a
build-dependency on redbpf, so that'd introduce a cycle;
- redbpf-probes can't depend on redbpf, since redbpf is not no_std.

At some point we'll probably have to create a redbpf-types crate or
something for the types we share between the two. Since there's only one
type at the moment, and it's not clear how many (if any) other types
we're going to have, at this time adding yet another crate isn't worth
the trouble.