redbpf: refactor API
The main change consists in removing ProgramKind and making Program an
enum with a variant for each probe type. In addition to being more
idiomatic, the new API makes it impossible to call the wrong attach_*
methods. With the old API it was in fact possible to call attach_xdp()
or attach_socketfilter() for a kprobe program, which was obviously
wrong.  With the new API that'd result in a compilation error.

redbpf::load::Loader was refactored too. The builder pattern is out and
the API now exports a similar interface for all program types. The
simplest kprobe loader now is something like:

        let mut loader = Loader::load_file("example.elf").unwrap();
        for kprobe in loader.kprobes_mut() {
            kprobe.attach_kprobe(&kprobe.name(), 0).unwrap();
        }

And the simplest XDP loader would be:

        let mut loader = Loader::load("example.elf").unwrap();
        for prog in loader.xdps_mut() {
            prog.attach_xdp("eth0", xdp::Flags::default()).unwrap();
        }