cargo-bpf: fix for rustc >= 1.45
Since version 1.45, rustc no longer produces bitcode for crates by
default. We need bitcode to turn on LTO, so pass -C embed-bitcode=yes to
the compiler.

We must pass it in RUSTFLAGS since we need bitcode for all crates, while
the flags passed as cargo rustc args only apply to the main crate.