Export `kernel_version` cfg variable (x, 5.7)
An ingraind the kprobe uses APIs that we know are changed after 5.7.

This is probably a bit rough this way, but here I export a
`kernel_version` cfg variable that is either the actual kernel
version, or `5.7` for kernels 5.7 or newer.

This may be a good way to also support older kernel releases that may
not have a particular feature.