checks rustc channel at compile time
This commit uses the `rustc_version` crate to check the Rust channel at
compile time and set `rustc` cfg flags depending appropriately. This
allows later checking if we are running on a nightly compiler or not,
and if so we can enable features that are nightly only (such as emitting
a compile time warning on use of deprecated proc macro).

Relates to #105