re-enables #[map("..")] syntax with deprecation
This commit re-enables the `#[map("..")]` syntax which links the map in
the `maps/..` section of the binary. This is for backwards compatibility
only, where the preferred syntax is `#[map]` or `#[map(link_section = "..")]`.

This commit also uses the code set up in e2a6c23 to determine if we're
on a nightly compiler channel, and if so enables the
`proc_macro::Diagnotic` API and which allows us to emit a compile time warning
about the deprecation of `#[map("..")]` if it is being used.

Relates to #105