doesnt require bindgen deps for lib use
This commit removes several deps from `bindgen` such as CLI, logging,
and `which` dependencies where there is no CLI of this tool of this
tool. Prior to this commit, `bindgen` includes `clap` and other crates
which then depend on several other crates adding compile time. When
using any of the `redbpf` crates purely as a library, there is no need
to pull in these additional crates.

This commit removes the default features from `bindgen`, and adds back
the only one that appears relevant to `redbpf` use (`runtime` feature).
However, if desired more default features could be added back, or the
`runtime` feature could be removed as well.

The default features from `bindgen` are:

- logging: Adds logging via the `log` crate
- clap: Adds a the `bindgen` CLI
- runtime: ensures `libclang` is loaded at runtime
- which-rustfmt: uses the `which` crate to determine the `rustfmt`
  binary location. Without this feature `bindgen` simply assumes
  `rustfmt` is somewhere in `$PATH`