Fixes zero sized ELF relocations
In some strange circumstances (I'm not sure exactly what triggers it
inside rustc) some sections will be generated with zero size. The loader
API was tripped up by trying to create a map with a value size of 0,
which will always fail verifier checks, since the value size must always
be a mutliple of the element requested (and nothing can be a multiple of
0).

Additionally, even if we don't create the map, the relocation will fail
if the size is 0 because we didn't create a map for it.

This commit fixes both issues, but only for section `.rodata`. If we
later determine that sections such as `.data`, or `.bss` exhibit similar
behavior we can add them to the `content.len() > 0` check.