Add an example of socket filter
This example describes how to write an eBPF program and a user-space
program which make use of socket filter.

It captures SYN packets and stores current timestamp into a map. When it
finds FIN or RST packets, it calculates TCP lifetime by subtracting the
timestamp stored at the map from the current timestamp.