Add Array and PerCpuArray structures for userspace
Array is simple structure wrapping methods of bpf_sys.

PerCpuArray should deal with per-cpu values and round up value size to
multiple of 8 bytes.

PerCpuValues is a companion structure which is passed to
PerCpuArray::set and returned by PerCpuArray::get method. It is a
newtype of Box<[T]> and the length of slice is the same with the number
of possible CPUs.