Use CARGO_ENCODED_RUSTFLAGS to pass --embed-bitcode=yes to rustc
Previously, cargo-bpf used the RUSTFLAGS environment variable to pass
--embed-bitcode=yes to all invocations of rustc. But since rust 1.55,
this behavior has changed. Even if RUSTFLAGS='-C embed-bitcode=yes' is
specified, this flag is not passed to the rustc so that compilation
fails with an error: "error: options `-C embed-bitcode=no` and `-C lto`
are incompatible".

Since cargo v1.55, this error occurs because it passes `-C
embed-bitcode=no` to rustc but RUSTFLAGS='-C embed-bitcode=yes' is not
appended to the rustc flags.

To resolve this problem, since rust v1.55 use CARGO_ENCODED_RUSTFLAGS
instead of RUSTFLAGS to pass --embed-bitcode=yes to rustc.

Signed-off-by: Junyeong Jeong <rhdxmr@gmail.com>