Fix cargo-bpf to build documentation in docs.rs
The documentation that is served by docs.rs is created by `cargo
rustdoc` command. And the command is executed inside a container that is
created by using the `ghcr.io/rust-lang/crates-build-env/linux:latest`
docker image.

For now the image is based on ubuntu 21.04 and it includes LLVM 10.
And when building documentations, nightly version of cargo is used.

In order to fix building cargo-bpf documentation, use llvm-sys-100 and
specify proper flags that are passed to cargo and rustdoc commands.

Signed-off-by: Junyeong Jeong <rhdxmr@gmail.com>