Fix building documentation of redbpf-probes in docs.rs
When building documentation of redbpf-probes in a container of docs.rs,
the DOCS_RS environment variable is set. If it is set, find all kernel
header directories in /lib/modules and pick one directory whose version
is the highest. And set KERNEL_SOURCE environment variable to the path.
This is conducted by the build script of redbpf-probes.

Signed-off-by: Junyeong Jeong <rhdxmr@gmail.com>