Move MapBtf structures to the maps.ext section
A MapBtf structure has been appended right after the bpf_map_def
structure. But this prohibits multiple maps from being defined in one
section. But the `maps` section should be able to accommodate multiple
maps. For example, all `TcHashMap`s should be defined in the `maps`
section. This restriction is required by tc utility.

Define all MapBtf structures in the `maps.ext` section so that multiple
`bpf_map_def` structures can be put together without interleaved MapBtf
structures.

Normally only one map is defined in a section. But defining more than
one maps in a section is semantically ambiguous. Because the name of the
section represents the name of the map. So sections of which name is
`maps/<name>` are not allowed to hold more than one bpf_map_def. Only
`maps` section is allowed to have multiple maps.

Signed-off-by: Junyeong Jeong <rhdxmr@gmail.com>