Add BTF::filter method
Filter particular BTF types that are chosen by a predicate closure. This
method is intended to be used by cargo-bpf to remove some BTF types that
are now supported by tc command.

If a BTF record is chosen to be deleted then all relevant BTF types that
depend on it are also eliminated first. So BTF consistency can be upheld.

Signed-off-by: Junyeong Jeong <rhdxmr@gmail.com>