Improve the echo example program that uses SockMap
Update sockmap with a new socket descriptor before any data is read from
the descriptor or before some epoll event about the descriptor
occurs. Because setting sockmap fails if those events happens before.

Read all remaining data from the socket and write it to the client to
serve the echo task. This happens when setting sockmap fails because the
connection is already half closed.

Signed-off-by: Junyeong Jeong <rhdxmr@gmail.com>