Add support for USDT semaphores
USDT probes can be used to define static trace points which
can provide helpful structured debug info. Due to performance
considerations, these probes can be disabled by default through
a mechanism called _semaphores_ which are basically pre-defined
memory locations which need to be rewritten to enable probes.
This can be done on a per-process basis by poking a given
memory address but you need to know a PID beforehand to do so.
The Linux kernel provides a mechanism to automatically override
semaphore addresses when you setup a uprobe.

This commit implements an optional argument for uprobes which can
be used to set a semaphore address to pass on to the kernel.

Signed-off-by: Nikita Baksalyar <nikita.baksalyar@gmail.com>