redbpf: compute symbol offset rather than address
BPF expects offsets in the object files rather than virtual memory addresses.
In some cases, the virtual memory offset of the .text section is 0, in which
case the correction in this change is a no-op. If the .text section has a
non-zero address offset, the previous logic for determining the symbol offset
was wrong.